modeling, material, and lighting changes
I created a lava material, changed the design of my main room so that the player can move around more. I created a prefab for a torch. I implemented some lighting as well. Lastly, I added a room where the second key will be (replaced the fountain with said room).