Design and Lighting modification + Doors and Climbing
I changed the design of the climbable areas of my level. I also added more lights to the beginning of the level. Lastly, I implemented some doors that open when the player is near and close when the player is not near. I will add onto this for my lock and key puzzle.